Apply Qodo /agentic_review pass 1: split event-arg files, non-blocking UI marshal, lease/discovery hardening
- Split BootloaderDiscoveredEventArgs and BootloaderHoldDroppedEventArgs into
  their own files (repo's single-public-type-per-file rule).
- BootloaderWatcher: marshal bound-collection mutations with the non-blocking
  Dispatcher.BeginInvoke instead of Invoke, so a UI mutation under _gate can't
  lock-invert with Dispose()'s synchronous _gate.Wait().
- FirmwareDialogViewModel: guard flashLease.DisposeAsync() in try/catch so a
  watcher-resume failure can't crash the command or mask the flash outcome.
- HidBootloaderDiscovery: Stop() now clears _loopTask so a Stop→Start resume
  restarts discovery immediately instead of no-opping until the old loop drains;
  the Core HidDeviceFinder is created via an injected factory (default = new).
- Wrap a >120-char log line in BootloaderHoldService.

Declined: tab indentation in the .csproj PackageReference line — the file is
tab-indented throughout and matching it is correct (same finding rejected in
PR #551/#499).

482 desktop tests green.
