+++ Daqifi.Desktop/Device/Firmware/IBootloaderDiscovery.cs
@@ -1,25 +1,6 @@
 namespace Daqifi.Desktop.Device.Firmware;
 
 /// <summary>
-/// Identifies a HID bootloader that discovery surfaced.
-/// </summary>
-public sealed class BootloaderDiscoveredEventArgs : EventArgs
-{
-    /// <summary>Creates the event args.</summary>
-    public BootloaderDiscoveredEventArgs(string devicePath, string? deviceName)
-    {
-        DevicePath = devicePath;
-        DeviceName = deviceName;
-    }
-
-    /// <summary>OS HID device path of the discovered bootloader.</summary>
-    public string DevicePath { get; }
-
-    /// <summary>Friendly device name, when discovery could read it; otherwise null.</summary>
-    public string? DeviceName { get; }
-}
-
-/// <summary>
 /// Continuous HID-bootloader discovery source for the <see cref="IBootloaderWatcher"/>. Abstracted from
 /// Core's <c>HidDeviceFinder</c> so the watcher can be unit-tested without hardware (tests raise
 /// <see cref="BootloaderDiscovered"/> directly) and so discovery can be paused around a flash.
